requirements.txt
py-cord
python-dotenv
asyncio
